Change: move first four filter functions out of manage_sql.c
