+++ src/gmp.c
@@ -106,6 +106,7 @@
 #include "manage_acl.h"
 #include "manage_alerts.h"
 #include "manage_assets.h"
+#include "manage_filters.h"
 #include "manage_oci_image_targets.h"
 #include "manage_port_lists.h"
 #include "manage_report_configs.h"
